feat: Add password field to Account interface and update signup functionality
This commit adds a password field to the Account interface in the Account.ts file. It also updates the signup functionality in the DataBaseService class to include the password when adding a new account to the database. This change allows for the storage and retrieval of account passwords, enhancing the security and functionality of the application.